Renomeia OperacoesDao e move pra pacote dal
